+++ docs/conf.py
@@ -124,11 +124,11 @@ pygments_style = 'friendly'
 
 autosummary_generate = True
 
-extlinks = {'issue': ('https://github.com/data-cube/agdc-v2/issues/%s', 'GH')}
+extlinks = {'issue': ('https://github.com/opendatacube/datacube-core/issues/%s', 'GH')}
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3.5/', None),
     'pandas': ('http://pandas.pydata.org/pandas-docs/stable/', None),
-    'numpy': ('http://docs.scipy.org/doc/numpy/', None),
+    'numpy': ('https://docs.scipy.org/doc/numpy/', None),
     'xarray': ('http://xarray.pydata.org/en/stable/', None),
     'dask': ('http://dask.pydata.org/en/stable/', None),
 }